+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support calling `/api/v1/accounts/search` with `resolve` parameter.
+
 ### Changed
 
 - Increase maximum number of custom emojis per post to 50.